requirements.txt
streamlit
pymongo
python-dotenv
bcrypt
scikit-learn
surprise
plotly
pandas
